# docs/swagger.yaml
definitions:
  repo.Article:
    properties:
      author:
        type: string
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      is_deleted:
        type: boolean
      likes:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    required:
    - author
    - content
    - id
    - title
    type: object
  repo.Author:
    properties:
      created_at:
        type: string
      full_name:
        type: string
      id:
        type: string
      password:
        type: string
      updated_at:
        type: string
      username:
        type: string
    required:
    - full_name
    - id
    - password
    - username
    type: object
info:
  contact: {}
paths:
  /v1/article/{id}:
    get:
      consumes:
      - application/json
      description: Get article by given ID.
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.Article'
      summary: get article by given ID
      tags:
      - Article
  /v1/author:
    post:
      consumes:
      - application/json
      description: Create author by given Body.
      parameters:
      - description: Author
        in: body
        name: author
        required: true
        schema:
          $ref: '#/definitions/repo.Author'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.Author'
      summary: get author by given ID
      tags:
      - Author
  /v1/author/{id}:
    get:
      consumes:
      - application/json
      description: Get author by given ID.
      parameters:
      - description: Author ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.Author'
      summary: get author by given ID
      tags:
      - Author
swagger: "2.0"